requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
influxdb-client==1.44.0
redis==5.0.1
paho-mqtt==1.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
urllib3==1.26.18
